requirements.txt
tweepy==4.14.0
praw==7.7.1
pandas==2.1.0
numpy==1.24.3
textblob==0.17.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
ccxt==4.1.13
matplotlib==3.7.2
seaborn==0.12.2
newsapi-python==0.2.7
pytrends==4.9.0
python-binance==1.0.19 